// src/common/verification/verification.enums.ts
export enum VerificationPurpose {
  PASSWORD_RESET = 'PASSWORD_RESET',
  CONTACT_VERIFICATION = 'CONTACT_VERIFICATION',
}

export enum VerificationSubjectType {
  USER = 'USER',
  EMERGENCY_CONTACT = 'EMERGENCY_CONTACT',
}

export enum VerificationChannel {
  EMAIL = 'EMAIL',
  WHATSAPP = 'WHATSAPP',
}